pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "socratic"
version = "0.1.0"
authors = [
  { name="Codeium", email="support@codeium.com" },
]
description = "A context-aware AI reasoning framework"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "dspy-ai>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "backoff>=2.2.1",
    "transitions>=0.9.0",
    "openai>=1.0.0"
]

[project.optional-dependencies]
test = ["pytest>=7.0.0"]
dev = ["black", "isort", "mypy"]

[tool.setuptools]
package-dir = {"" = "socratic"}

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3
